+++ app/[locale]/login/page.tsx
@@ -1,4 +1,4 @@
-import { signIn } from "@/auth";
+import { signIn } from "@/utils/auth";
 import LoginForm from "./components/LoginForm";
 import { AuthError } from "next-auth";
 import { IResponse } from "@/types";
