+++ app/api/users/is-username-unique/[username]/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import connectDB from "@/lib/mongodb";
+import connectDB from "@/utils/mongodb";
 import User from "../../../models/users-model";
 import { IResponse } from "@/types";
