+++ middleware.ts
@@ -2,7 +2,7 @@ import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
 import createMiddleware from "next-intl/middleware";
 import { routing } from "./i18n/routing";
-import { auth } from "@/auth";
+import { auth } from "@/utils/auth";
 
 export async function middleware(request: NextRequest) {
   const intlMiddleware = createMiddleware(routing);
